fix(api): cast exchange rate to float in CurrencyConversionService
fix(web): display listing title in dashboard bookings
fix(web): correct price formatting (remove /100 division)

+++ apps/laravel-api/app/Services/CurrencyConversionService.php
@@ -56,7 +56,7 @@ class CurrencyConversionService
 
         $cacheKey = "exchange_rate:{$fromCurrency}_{$toCurrency}";
 
-        return Cache::remember($cacheKey, now()->addHours(1), function () use ($fromCurrency, $toCurrency) {
+        return (float) Cache::remember($cacheKey, now()->addHours(1), function () use ($fromCurrency, $toCurrency) {
             if ($fromCurrency === self::BASE_CURRENCY) {
                 // Direct conversion from TND
                 $rate = ExchangeRate::where('currency', $toCurrency)
